requirements.txt
Django==3.1.3
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
parameterized==0.7.4

